Renamed CreateCacheItem() to NewCacheItem().
The major difference being NewCacheItem now returns a pointer to a
cache item. This prevents copying the mutex internally, which is
nasty. Additionally, all methods for CacheItem are pointer receivers
already by design, so this fits better.

DRY: Added addInternal() to remove duplicated code.

This will break API with previous versions of cache2go. I've tagged
the old API with 'v0.1' in git.